+++ mopidy/frontends/mpd/server.py
@@ -29,14 +29,14 @@ class MpdServer(asyncore.dispatcher):
             else:
                 self.create_socket(socket.AF_INET, socket.SOCK_STREAM)
             self.set_reuse_addr()
-            hostname = self._format_hostname(settings.SERVER_HOSTNAME)
-            port = settings.SERVER_PORT
+            hostname = self._format_hostname(settings.MPD_SERVER_HOSTNAME)
+            port = settings.MPD_SERVER_PORT
             logger.debug(u'Binding to [%s]:%s', hostname, port)
             self.bind((hostname, port))
             self.listen(1)
             logger.info(u'MPD server running at [%s]:%s',
-                self._format_hostname(settings.SERVER_HOSTNAME),
-                settings.SERVER_PORT)
+                self._format_hostname(settings.MPD_SERVER_HOSTNAME),
+                settings.MPD_SERVER_PORT)
         except IOError, e:
             sys.exit('MPD server startup failed: %s' % e)