+++ docs/changes.rst
@@ -17,6 +17,8 @@ Another great release.
 - Include Sphinx scripts for building docs, pylintrc, tests and test data in
   the packages created by ``setup.py`` for i.e. PyPI.
 - Rename :mod:`mopidy.backends.gstreamer` to :mod:`mopidy.backends.local`.
+- Changed ``SERVER_HOSTNAME`` and ``SERVER_PORT`` settings to
+  ``MPD_SERVER_HOSTNAME`` and ``MPD_SERVER_PORT``
 - MPD frontend:
 
   - Relocate from :mod:`mopidy.mpd` to :mod:`mopidy.frontends.mpd`.